[RENAME] : sidebar.tsx 위치 변경으로 import 수정
