+++ src/pages/Lost/views/Lost.tsx
@@ -1,23 +1,23 @@
 import React from 'react';
 import LostDoc from './LostDoc';
-import SideBar from '../../../components/SideBar/SideBar';
+import SideBar from '../../../components/SideBar/views/SideBar';
 
 function Lost() {
 	return (
-        <>
-        <SideBar />
-        <div className='bg-nomad-sand min-h-full pt-9'>
-            <div className='grid grid-cols-2 gap-5 p-8 box-border'>
-                <LostDoc />
-                <LostDoc />
-                <LostDoc />
-                <LostDoc />
-                <LostDoc />
-                <LostDoc />
-            </div>
-        </div>
-        </>
-    )
+		<>
+			<SideBar />
+			<div className="bg-nomad-sand min-h-full pt-9">
+				<div className="grid grid-cols-2 gap-5 p-8 box-border">
+					<LostDoc />
+					<LostDoc />
+					<LostDoc />
+					<LostDoc />
+					<LostDoc />
+					<LostDoc />
+				</div>
+			</div>
+		</>
+	);
 }
 
 export default Lost;
