+++ src/pages/Main/views/Main.tsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import SideBar from '../../../components/SideBar/SideBar';
+import SideBar from '../../../components/SideBar/views/SideBar';
 
 function Main() {
 	return (
